fix: onCreateなどを整理
