+++ functions/src/schedules/contract-normal.spec.ts
