+++ functions/src/schedules/contract-renewable.ts
@@ -1,9 +1,9 @@
 /* eslint-disable camelcase */
-import { single_price_renewable_settlement } from '.';
 import { renewable_ask_history } from '../renewable-ask-histories';
 import { renewable_ask } from '../renewable-asks';
 import { renewable_bid_history } from '../renewable-bid-histories';
 import { renewable_bid } from '../renewable-bids';
+import { single_price_renewable_settlement } from '../single-price-renewable-settlements';
 import { RenewableAskHistory, RenewableBidHistory, SinglePriceRenewableSettlement } from '@local/common';
 import * as functions from 'firebase-functions';
