+++ functions/src/schedules/operation-normal.ts
@@ -1,9 +1,9 @@
 /* eslint-disable camelcase */
-import { normal_ask } from '.';
 import { admin_account } from '../admin-accounts';
 import { delta_amount } from '../delta-amounts';
 import { normal_ask_history } from '../normal-ask-histories';
 import { normal_ask_setting } from '../normal-ask-settings';
+import { normal_ask } from '../normal-asks';
 import { normal_bid_history } from '../normal-bid-histories';
 import { normal_bid } from '../normal-bids';
 import { single_price_normal_settlement } from '../single-price-normal-settlements';
